refactor: move utils.ts to utils/index.ts for coverage gate compatibility
Rename utils.ts → utils/index.ts in k8s, docker, and hooklib packages so
source files match the CI coverage gate pattern '**/utils/**'. Add
collectCoverageFrom config to jest.config.js in k8s and docker packages to
enable coverage reporting. All existing imports resolve unchanged via TS
module resolution.
